# storage/api-docs/api-docs.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "LeSGo API",
    "version": "1.0.0",
    "description": "Logistics & multi-service API for LeSGo — ride, delivery, buy, and eat services.",
    "contact": {
      "email": "support@lesgo.app"
    }
  },
  "servers": [
    {
      "url": "http://localhost:8000",
      "description": "Local Development Server"
    },
    {
      "url": "https://lesgo-api-feature-auth-secmes.free.laravel.cloud",
      "description": "Staging Server"
    }
  ],
  "components": {
    "securitySchemes": {
      "sanctum": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Enter your Sanctum token: Bearer {token}"
      }
    },
    "schemas": {
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": false
          },
          "message": {
            "type": "string",
            "example": "Error message"
          },
          "request_id": {
            "type": "string",
            "example": "550e8400-e29b-41d4-a716-446655440000"
          }
        }
      }
    }
  },
  "paths": {
    "/api/v1/ping": {
      "get": {
        "summary": "Health check",
        "tags": ["System"],
        "responses": {
          "200": {
            "description": "API is up",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "LeSGo API v1 OK"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/vouchers/available": {
      "get": {
        "summary": "Get available vouchers for current user",
        "tags": ["Vouchers"],
        "security": [{"sanctum": []}],
        "responses": {
          "200": {
            "description": "Available vouchers",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {"type": "boolean", "example": true},
                    "data": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "code": {"type": "string"},
                          "title": {"type": "string"},
                          "description": {"type": "string"},
                          "eligible": {"type": "boolean"}
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/vouchers/validate": {
      "post": {
        "summary": "Validate voucher code for current user",
        "tags": ["Vouchers"],
        "security": [{"sanctum": []}],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["voucher_code", "order_value"],
                "properties": {
                  "voucher_code": {"type": "string", "example": "WELCOME10"},
                  "order_value": {"type": "number", "format": "float", "example": 150.00}
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Voucher validation result",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {"type": "boolean", "example": true},
                    "data": {
                      "type": "object",
                      "properties": {
                        "valid": {"type": "boolean"},
                        "discount_amount": {"type": "number", "format": "float"},
                        "discount_type": {"type": "string"},
                        "new_total": {"type": "number", "format": "float"}
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    }
  }
}